requirements.txt
python-hcl2==4.3.2
rich==13.7.1

